- Added TAGS file generated using the following bash command: rm -f TAGS; for f in `find . -iname '*.lisp'`; do etags --append "$f"; done
 - Renamed the macro-magic directory to tools/

 - Moved several books from misc/ into tools/.

NOTE: This branch will probably not currently pass regression testing.
Certain further rearrangements may be necessary in order to make it
possible to build with the current Makefile system due to circular
dependencies among directories.

NOTE2: This could be considered a fairly drastic change since many of
the books that were moved are probably in common use; hence the
separate branch.  Before merging this to the trunk, I hope to hash out
which books really should have been moved (and, obviously, get it to
pass regression.)


git-svn-id: http://acl2-books.googlecode.com/svn/branches/misc-cleanup@20 303c5b2c-9740-0410-b014-fd8f9db7be32