Add new std/alists directory.  For now this is pretty minimal.
Replace some of vl/util/arith.lisp with the new std/alists stuff.

Add cutil/defruled, which is to defrule as defthmd is to defthm.

Localized a couple of general-purpose alist lemmas in witness-cp, since they
aren't really what the book is about.

Slight revisions of vl-lint, adding some minor features and documentation.


git-svn-id: http://acl2-books.googlecode.com/svn/trunk@1665 303c5b2c-9740-0410-b014-fd8f9db7be32