XDOC now has a new, "fancy" viewer.
   Preview: http://fv.centtech.com/fancy/    (slow loading, see below)
     (Note: will be taken down soon)

The basic idea is that instead of generating everything (index and content
pages) ahead of time as thousands of individual .xml and .xdoc-link files, I
just preprocess everything and then write out the XDOC database as JSON
formatted data.

On the ACL2 side, this approach simplifies the XDOC lisp code by a good deal
(well---or at least, it *could* simplify the code, if I got rid of the old
"save" command, but I've left that in for now).  It also speeds up creating
manuals by a good bit (it cuts about a minute off of centaur/doc.lisp; with the
actual save command is down to about 25 seconds.)

On the display side, things are more complex but much fancier.  The JSON
version of the documentation database can be easily loaded into JavaScript, so
the entire display can be handled with ordinary web programming.  The new
viewer has a much richer navigation system with both hierarchical and flat
organizations, A-Z buttons for the flat navigation, a "jump to topic" box with
typeahead support, etc.  It also has a much fancier content display that lets
you expand subtopics, etc.  This is really cool: try it for instance on

       http://fv.centtech.com/fancy/?topic=ACL2____RULESETS

Deployment is somewhat easier than before.  There are far fewer files and no
need to install Xalan or run "make" any more.

One problem is size.  The encoded database is large (around 20 MB), and we just
load the whole stupid thing in.  This is just fine (great, even) when reading
from local storage or from an intranet site on a LAN.  But it's definitely not
a good solution for public facing web sites like the fv.centtech.com above, and
external users will probably see a considerable delay while the page boots up.

For a public-facing web site, it should be quite easy to stuff the big xdata.js
part of the database into, say, an sqlite3 database and grab the long strings
on demand via AJAX queries.  I'll probably add this soon to make a much faster
version of the fancy web site.




git-svn-id: http://acl2-books.googlecode.com/svn/trunk@1945 303c5b2c-9740-0410-b014-fd8f9db7be32