Update Quicklisp from 2014-01-13 to 2014-06-16 and remove dependency on iolib
Some day we might revive a connection with iolib.  For now, though, its build
process doesn't seem very stable; see for instance:

    https://groups.google.com/forum/#!topic/quicklisp/B1byNhwmUtU

So far we've avoided inflicting this problem on folks by just using an older
version of Quicklisp.

However, it turned out that iolib was being used only for getpid, which happens
to be available in ACL2 as the getpid$ function.  So for now, it seems easiest
to just get rid of iolib and use ACL2's version.

git-svn-id: http://acl2-books.googlecode.com/svn/trunk@2824 303c5b2c-9740-0410-b014-fd8f9db7be32