Added license/copyright info after getting the OK from relevant folks.
In consultation with Jared Davis and David Rager, moved the definition
of non-parallel-book from centaur/vl/util/defs.lisp to
std/system/non-parallel-book.lisp, and included that in
std/system/top, now included in books/doc/top.lisp.  Used
non-parallel-book in a couple of new places:

  centaur/bitops/equal-by-logbitp.lisp
  centaur/bitops/congruences.lisp

Also enhanced the definition of non-parallel-book with keyword
arguments (not yet used) and extended its documentation, and changed
its documentation parent to parallelism.

NOTE: In support of the above, added the symbol
acl2::non-parallel-book to the "VL" package in centaur/vl/package.lsp.


git-svn-id: http://acl2-books.googlecode.com/svn/trunk@2889 303c5b2c-9740-0410-b014-fd8f9db7be32